chore : 실제 휴대폰 연동 수정(#1)

+++ app/src/main/java/com/example/moing/Response/InviteTeamResponse.java
@@ -26,9 +26,9 @@ public class InviteTeamResponse {
 
     public static class Data {
         @SerializedName("teamId")
-        private int teamId;
+        private Long teamId;
 
-        public int getTeamId() {
+        public Long getTeamId() {
             return teamId;
         }
     }
